requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
chromadb==0.5.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
loguru==0.7.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pyyaml==6.0.1
rdflib==7.0.0
requests==2.32.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
streamlit==1.37.0
toml==0.10.2
unmarkd==1.1.3
urllib3==2.2.1